Moving dilutions classes to its own file
